Fetch preference data
