+++ src/applications/personalization/dashboard/dashboard-entry.jsx
@@ -8,7 +8,7 @@ import '../preferences/sass/preferences.scss';
 import startApp from 'platform/startup';
 
 import routes from './routes';
-import reducer from './reducers/';
+import reducer from './reducers';
 import manifest from './manifest';
 
 startApp({
