+++ src/applications/personalization/preferences/components/PreferenceList.jsx
@@ -14,7 +14,7 @@ export default function PreferenceList({
         <PreferenceItem
           key={index}
           benefit={benefit}
-          isRemoving={view[benefit.slug]}
+          isRemoving={view[benefit.code]}
           handleViewToggle={handleViewToggle}
           handleRemove={handleRemove}
         />
