+++ src/applications/personalization/preferences/components/PreferenceOption.jsx
@@ -4,14 +4,14 @@ import Checkbox from '../../../gi/components/Checkbox.jsx';
 
 export default ({ item, onChange, checked }) => {
   const itemContent = (
-    <div className="preference-item" onClick={onChange(item.code)}>
+    <div className="preference-item" onClick={() => onChange(item.code)}>
       <div className="title-container">
         <h5 className="title-item">{item.title}</h5>
         <Checkbox
           name={item.code}
           checked={checked}
           label=""
-          onChange={onChange(item.code)}
+          onChange={() => onChange(item.code)}
         />
       </div>
       <p>{item.description}</p>
