+++ src/applications/personalization/preferences/containers/PreferencesWidget.jsx
@@ -15,7 +15,7 @@ import {
   setPreference,
   savePreferences,
   fetchUserSelectedBenefits,
-} from '../actions/actions';
+} from '../actions';
 import { benefitChoices } from '../helpers';
 import { LOADING_STATES } from '../constants';
